+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.2] - unreleased
+## [0.2] - 2026-06-28
 
 ### Added